+++ ind.css
@@ -1,0 +1,5 @@
+s {
+	color: #0c381e;
+	color: red;
+	color: red;
+}
