+++ plugin/color_picker/linux_executable.py
