+++ plugin/color_picker/win_colorpicker.exe
